Fix rerenderin home bug
